Basic enemy waves
Upon entering a unvisited room (that has enemy spawners in it), all doors will be closed and enemies will spawn. Refactoring and clean up required, but as a first concept it's fine.
Further development is required in the enemy script so that the enemy reports back when it died to the spawner (when we have enemies).