Added Active buff-items, following structure and groundwork for items.
Note: Some groundwork or foundation still missing for different item behaviours. 

Push adds 4 scripts, 2 of which are scriptable objects (Active, A_Buff), the third being the CooldownController that the player prefab now uses to maintain it's current Active object and it's usage + cooldowns. Fourth script is a simple itemholder list. Also added Itemholder to gamecontroller prefab to hold items (this will later be used when player needs to interract with an item to pick it up, probably..)